+++ packages/editor/src/ce/components/document-editor-side-effects.ts
