+++ packages/editor/src/core/components/editors/document/collaborative-editor.tsx
@@ -13,7 +13,7 @@ import { getEditorClassNames } from "@/helpers/common";
 // hooks
 import { useCollaborativeEditor } from "@/hooks/use-collaborative-editor";
 // constants
-import { DocumentEditorSideEffects } from "@/plane-editor/components/editor-side-effects";
+import { DocumentEditorSideEffects } from "@/plane-editor/components/document-editor-side-effects";
 // types
 import type { EditorRefApi, ICollaborativeDocumentEditorProps } from "@/types";
